chap4 summit

+++ chap4/prob7/student.h
@@ -1,5 +1,5 @@
-struct student{
+struct student{	
 	int id;
 	char name[20];
-	int score;
+	short score;
 };
